more store cleanup
